move DrivesResponseError class
