+++ src/handler.ts
@@ -2,7 +2,7 @@ import { URLExt } from '@jupyterlab/coreutils';
 import { ServerConnection } from '@jupyterlab/services';
 import { ReadonlyJSONObject } from '@lumino/coreutils';
 
-import { DrivesResponseError } from './drivesError';
+// import { DrivesResponseError } from './drivesError';
 
 /**
  * Array of Jupyter Drives Auth Error Messages.
@@ -14,6 +14,39 @@ export const AUTH_ERROR_MESSAGES = [
   'could not read Session Token',
   'Authentication error'
 ];
+
+/**
+ * A wrapped error for a fetch response.
+ */
+export class DrivesResponseError extends ServerConnection.ResponseError {
+  /**
+   * Create a new response error.
+   */
+  constructor(
+    response: Response,
+    message = `Invalid response: ${response.status} ${response.statusText}`,
+    traceback = '',
+    json: ReadonlyJSONObject | null = {}
+  ) {
+    super(response, message);
+    this.traceback = traceback; // traceback added in mother class in 2.2.x
+    this._json = json ?? {};
+  }
+
+  /**
+   * The error response JSON body
+   */
+  get json(): ReadonlyJSONObject {
+    return this._json;
+  }
+
+  /**
+   * The traceback associated with the error.
+   */
+  traceback: string;
+
+  protected _json: ReadonlyJSONObject;
+}
 
 /**
  * Call the API extension
